requirements.txt
# EtherScope - Web3 Wallet Intelligence Telegram Bot
# Production dependencies

# Core dependencies
python-dotenv==1.0.0
python-telegram-bot==20.7
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0

# Async support
aiohttp==3.9.1

# Optional: Analytics
pandas==2.1.3
numpy==1.26.2

# Development dependencies (for development only)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
mypy==1.7.1
isort==5.13.2

# Production server (optional, for deployment)
gunicorn==21.2.0
